# specs/001-todo-fullstack-app/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Todo Full-Stack API
  description: |
    RESTful API for the Todo Full-Stack Web Application.
    All endpoints require JWT authentication via Better Auth.
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server

security:
  - BearerAuth: []

tags:
  - name: Tasks
    description: Task CRUD operations
  - name: Health
    description: Health check endpoints

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      description: Returns API health status (no auth required)
      security: []
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  timestamp:
                    type: string
                    format: date-time

  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: List user's tasks
      description: |
        Returns paginated list of tasks owned by the authenticated user.
        Tasks are sorted by creation date (newest first).
      operationId: listTasks
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Items per page
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed]
          description: Filter by task status
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Tasks]
      summary: Create a new task
      description: Creates a new task owned by the authenticated user
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
        description: Task ID

    get:
      tags: [Tasks]
      summary: Get a single task
      description: Returns a task if owned by the authenticated user
      operationId: getTask
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tasks]
      summary: Update a task
      description: Updates a task if owned by the authenticated user
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: Delete a task
      description: Permanently deletes a task if owned by the authenticated user
      operationId: deleteTask
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/v1/tasks/{taskId}/complete:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: integer
        description: Task ID

    patch:
      tags: [Tasks]
      summary: Toggle task completion status
      description: Toggles task between pending and completed status
      operationId: toggleTaskComplete
      responses:
        '200':
          description: Task status toggled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Better Auth.
        Include as: Authorization: Bearer <token>

  schemas:
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title (required)
          example: Buy groceries
        due_date:
          type: string
          format: date
          nullable: true
          description: Optional due date (YYYY-MM-DD)
          example: "2025-01-15"

    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title
          example: Buy groceries and cook dinner
        due_date:
          type: string
          format: date
          nullable: true
          description: Due date (YYYY-MM-DD)
          example: "2025-01-20"
        status:
          type: string
          enum: [pending, completed]
          description: Task status
          example: completed

    TaskResponse:
      type: object
      required:
        - id
        - title
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Task ID
          example: 1
        title:
          type: string
          description: Task title
          example: Buy groceries
        due_date:
          type: string
          format: date
          nullable: true
          description: Due date
          example: "2025-01-15"
        status:
          type: string
          enum: [pending, completed]
          description: Task status
          example: pending
        created_at:
          type: string
          format: date-time
          description: Creation timestamp (UTC)
          example: "2025-01-01T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp (UTC)
          example: "2025-01-01T10:00:00Z"

    TaskListResponse:
      type: object
      required:
        - tasks
        - total
        - page
        - per_page
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
        total:
          type: integer
          description: Total number of tasks
          example: 42
        page:
          type: integer
          description: Current page number
          example: 1
        per_page:
          type: integer
          description: Items per page
          example: 20

    Error:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: Task not found

  responses:
    BadRequest:
      description: Bad Request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Title is required"

    Unauthorized:
      description: Unauthorized - Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Missing authorization header"

    Forbidden:
      description: Forbidden - User does not own this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "You do not have permission to access this task"

    NotFound:
      description: Not Found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            detail: "Task not found"